requirements.txt
aiomysql==0.2.0
albucore==0.0.24
albumentations==2.0.6
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
babel==2.17.0
bcrypt==4.3.0
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
Cython==3.0.12
dnspython==2.7.0
docxcompose==1.4.0
docxtpl==0.19.1
easydict==1.13
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.11
flatbuffers==25.2.10
fonttools==4.57.0
geographiclib==2.0
geopy==2.4.1
greenlet==3.1.1
h11==0.14.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
insightface==0.7.3
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
kombu==5.5.3
lazy_loader==0.4
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
onnx==1.17.0
onnxruntime==1.21.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
passlib==1.7.4
pillow==11.2.1
prettytable==3.16.0
prompt_toolkit==3.0.51
protobuf==6.30.2
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
qrcode==8.1
redis==5.3.0
requests==2.32.3
rsa==4.9
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
stringzilla==3.12.5
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.3.30
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
